refactor: parser클래스 삭제 및 메서드 이동
